janny/commandline: moved to own file for cleanliness
